moved next packet sequence query test
